headwind-mdm-maps-lite-main/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
psycopg2-binary==2.9.9
bcrypt==4.1.2
python-dotenv==1.0.0
